fix: TT-401 change in activity database instance and refactor test
