+++ V2/time_tracker/activities/_infrastructure/_data_persistence/__init__.py
@@ -1,2 +1,2 @@
 # flake8: noqa
-from ._activities_sql_dao import ActivitiesSQLDao
+from ._activities_dao import ActivitiesSQLDao
