+++ V2/time_tracker/time_entries/_application/_time_entries/_create_time_entry.py
@@ -56,6 +56,7 @@ def create_time_entry(req: func.HttpRequest) -> func.HttpResponse:
 def _validate_time_entry(time_entry_data: dict) -> typing.List[str]:
     time_entry_fields = [field.name for field in dataclasses.fields(_domain.TimeEntry)]
     time_entry_fields.pop(8)
+    time_entry_fields.pop(0)
     missing_keys = [field for field in time_entry_fields if field not in time_entry_data]
     return [
         f'The {missing_key} key is missing in the input data'
